When the user clicks unfollow it decreases the population count.
The primary key of SubFollower is user and sub. If a subfollower object can be obtained i.e not null, then the subFollower can be removed from the database and the population can be decreased.

+++ src/main/java/com/jon/attachat/controller/SubController.java
@@ -101,9 +101,13 @@ public class SubController {
 		
 		User user = userService.getUser(request.getUserPrincipal().getName());
 		Sub sub = subService.getSub(subName);
-		SubFollower subFollower = new SubFollower(user, sub);
-
-		subService.removeFollower(subFollower);
+		SubFollowerId id = new SubFollowerId(user, sub);
+		SubFollower subFollower = subService.getSubFollower(id);
+		
+		if(subFollower != null) {
+			subService.removeFollower(subFollower);
+			subService.decreaseSubPopulationCount(sub);
+		}
 		
 		redirectAttributes.addAttribute("subName", subName);
